[FIX] repairs warnings in Network package
remove unused parameters
fix conversion ID variable

+++ include/Network/Client.h
@@ -80,13 +80,13 @@ public:
 	void lookAt(osg::Vec3d coord);
 
 	void sendNewNode(QString name, osg::Vec3f position);
-	void sendNewEdge(QString name, int id_from, int id_to, bool oriented = false);
+	void sendNewEdge(QString name, qlonglong id_from, qlonglong id_to, bool oriented = false);
 	void sendRemoveNode(qlonglong id);
-	void sendRemoveEdge(int id);
+	void sendRemoveEdge(qlonglong id);
 	void sendNodeColor(qlonglong id, float r, float g, float b, float alpha);
 	void sendEdgeColor(int id, float r, float g, float b, float alpha);
-	void sendNodeLabel(int id, QString label);
-	void sendFixNodeState(int id, bool state);
+	void sendNodeLabel(qlonglong id, QString label);
+	void sendFixNodeState(qlonglong id, bool state);
 	void sendMergeNodes(QLinkedList<osg::ref_ptr<Data::Node> > * selectedNodes, osg::Vec3f position);
 	void sendSeparateNodes(QLinkedList<osg::ref_ptr<Data::Node> > * selectedNodes);
 	void sendAddMetaNode(QString name, QLinkedList<osg::ref_ptr<Data::Node> > * selectedNodes, QString edgeName, osg::Vec3f position);
